feat: add user and auth service

+++ src/services/user.service.ts
@@ -1,5 +1,6 @@
 import { User } from "../models";
 import { hashPassword } from "../utilities/hashPassword";
+import { CreateUser, UpdateUser } from "../types";
 
 export class UserService {
   static async getUsers() {
@@ -14,7 +15,7 @@ export class UserService {
     return User.findOne({ email });
   }
 
-  static async createUser(user: any) {
+  static async createUser(user: CreateUser) {
     const userWithSameEmail = await this.getUserByEmail(user.email);
 
     if (userWithSameEmail) {
@@ -34,11 +35,31 @@ export class UserService {
     return User.create(user);
   }
 
-  static async updateUser(userId: string, user: any) {
+  static async updateUser(userId: string, user: UpdateUser) {
+    const userExists = await this.getUserById(userId);
+
+    if (!userExists) {
+      throw new Error("User not found");
+    }
+
+    const userWithSameUsername = await User.findOne({
+      username: user.username,
+    });
+
+    if (userWithSameUsername) {
+      throw new Error("Username already in use");
+    }
+
     return User.findByIdAndUpdate(userId, user, { new: true });
   }
 
   static async deleteUser(userId: string) {
+    const userExists = await this.getUserById(userId);
+
+    if (!userExists) {
+      throw new Error("User not found");
+    }
+
     return User.findByIdAndDelete(userId);
   }
 }
